Refactor the code in RecordFieldCompletionUtil
